process-via-/proc: split code across a JSON-based API; enable unit tests
Also implements a GetProcessStartTime()/GetProcessState() API within that code,
eliminating (almost) "process_/OS/.c" when that OS uses the "/proc" mechanism.
The one residual exception is for the "linux_process_test" unit test
which needs deeper work to go through the new code.
